baseline_checker.py
#!/usr/bin/env python3
"""
baseline_checker.py

Assess a GitHub repo against ALL Level 1 controls of
the OpenSSF OSPS Security Baseline v2025-02-25.
Requires: PyGithub, click
Set your token in GITHUB_TOKEN or pass via --token.
"""

import os
import sys
from fnmatch import fnmatch
import click
from github import Github, GithubException

# ─── Helpers ───────────────────────────────────────────────────────────────────

def check_mfa(repo):
    # OSPS-AC-01.01
    owner = repo.owner
    if owner.type == "Organization":
        org = repo._requester.requestJsonAndCheck("GET", f"/orgs/{owner.login}")[1]
        return "PASS" if org.get("two_factor_requirement_enabled") else "FAIL"
    return "UNKNOWN (not an org)"

def check_default_permissions(repo):
    # OSPS-AC-02.01
    owner = repo.owner
    if owner.type == "Organization":
        org = repo._requester.requestJsonAndCheck("GET", f"/orgs/{owner.login}")[1]
        default = org.get("default_repository_permission")
        return "PASS" if default in ("read", "triage") else f"FAIL (default={default})"
    return "UNKNOWN (not an org)"

def check_branch_protection(repo):
    # OSPS-AC-03.01 & 03.02
    try:
        repo.get_branch(repo.default_branch).get_protection()
        return "PASS"
    except GithubException as e:
        if e.status == 404:
            return "FAIL (no protection)"
        return f"ERROR ({e.status})"

def list_files(repo, path=""):
    """Recursively list all files under `path`."""
    files = []
    try:
        contents = repo.get_contents(path)
    except GithubException:
        return files
    for entry in contents:
        if entry.type == "dir":
            files += list_files(repo, entry.path)
        else:
            files.append(entry.path)
    return files

def exists_any(repo, paths):
    for p in paths:
        try:
            repo.get_contents(p)
            return True
        except GithubException as e:
            if e.status != 404:
                return None  # error
    return False

def check_sanitized_pipeline(repo):
    # OSPS-BR-01.01: placeholder—detect any GitHub Actions workflow
    wf = exists_any(repo, ["*.yml", "*.yaml"] if False else [f".github/workflows/{f}" for f in ["*"]])
    if wf is None: return "ERROR scanning workflows"
    return "PASS" if list_files(repo, ".github/workflows") else "FAIL (no workflows)"

def check_encrypted_channel(repo):
    # OSPS-BR-03.01: homepage URL must be HTTPS
    url = repo.homepage or ""
    return "PASS" if url.startswith("https://") else f"FAIL (homepage={url or 'unset'})"

def check_user_docs(repo):
    # OSPS-DO-01.01: user guides
    return "PASS" if exists_any(repo, [
        "docs/USER_GUIDE.md", "docs/Getting_Started.md",
        "docs/getting-started.md", "USER_GUIDE.md", "getting_started.md"
    ]) else "FAIL"

def check_defect_reporting_guide(repo):
    # OSPS-DO-02.01: guide for reporting defects
    # check docs or ISSUE_TEMPLATE
    if exists_any(repo, [
        "docs/REPORTING_DEFECTS.md", "docs/bug_report.md"
    ]):
        return "PASS"
    # check GitHub issue template
    templates = list_files(repo, ".github/ISSUE_TEMPLATE")
    if any(fnmatch(t, "*bug*") for t in templates):
        return "PASS"
    return "FAIL"

def check_public_discussions(repo):
    # OSPS-GV-02.01
    return "PASS" if repo.has_issues or repo.has_wiki else "FAIL"

def check_contributing_guide(repo):
    # OSPS-GV-03.01
    return "PASS" if exists_any(repo, ["CONTRIBUTING.md", "CONTRIBUTING.MD", "CONTRIBUTING/"]) else "FAIL"

def check_license_spdx(repo):
    # OSPS-LE-02.01
    try:
        lic = repo.get_license().license
        spdx = lic.spdx_id if lic else None
        return "PASS" if spdx and spdx != "NOASSERTION" else f"FAIL (spdx={spdx})"
    except GithubException:
        return "FAIL (no license API)"

def check_license_assets(repo):
    # OSPS-LE-02.02 & 03.02: approximate via license file
    return "PASS" if exists_any(repo, ["LICENSE", "LICENSE.md", "COPYING"]) else "FAIL"

def check_license_file(repo):
    # OSPS-LE-03.01
    return check_license_assets(repo)

def check_repo_public(repo):
    # OSPS-QA-01.01
    return "PASS" if not repo.private else "FAIL"

def check_change_history(repo):
    # OSPS-QA-01.02
    try:
        cnt = repo.get_commits().totalCount
        return "PASS" if cnt > 0 else "FAIL (no commits)"
    except GithubException:
        return "ERROR fetching commits"

def check_dependency_list(repo):
    # OSPS-QA-02.01: common dependency files
    patterns = [
        "requirements.txt", "Pipfile.lock", "environment.yml",
        "package.json", "pom.xml", "build.gradle", "go.mod",
        "Gemfile", "composer.json"
    ]
    return "PASS" if exists_any(repo, patterns) else "FAIL"

def check_subprojects_list(repo):
    # OSPS-QA-04.01
    return "PASS" if exists_any(repo, [
        "SUBPROJECTS.md", "SUBPROJECTS.MD", "CODEBASES.md",
        "docs/SUBPROJECTS.md"
    ]) else "FAIL"

def check_no_binary_artifacts(repo):
    # OSPS-QA-05.01
    bins = [f for f in list_files(repo) if fnmatch(f, "*.exe") or fnmatch(f, "*.dll")
            or fnmatch(f, "*.bin") or fnmatch(f, "*.class") or fnmatch(f, "*.jar")]
    return "FAIL (found binaries)" if bins else "PASS"

def check_security_contacts(repo):
    # OSPS-VM-02.01
    return "PASS" if exists_any(repo, ["SECURITY.md", ".github/SECURITY.md"]) else "FAIL"

# ─── CLI ──────────────────────────────────────────────────────────────────────

@click.command()
@click.option("--repo", required=True, help="owner/name, e.g. openssf/baseline")
@click.option("--token", envvar="GITHUB_TOKEN", help="GitHub token")
def main(repo, token):
    if not token:
        click.echo("❌ Missing GitHub token (set GITHUB_TOKEN or use --token)")
        sys.exit(1)
    gh = Github(token)
    try:
        r = gh.get_repo(repo)
    except Exception as e:
        click.echo(f"❌ Could not open {repo}: {e}")
        sys.exit(1)

    checks = [
        ("OSPS-AC-01.01 MFA enforcement",            check_mfa(r)),
        ("OSPS-AC-02.01 Default collaborator perms", check_default_permissions(r)),
        ("OSPS-AC-03.01 Prevent direct commits",     check_branch_protection(r)),
        ("OSPS-AC-03.02 Prevent branch deletion",    check_branch_protection(r)),
        ("OSPS-BR-01.01 Pipeline sanitization",      check_sanitized_pipeline(r)),
        ("OSPS-BR-03.01 Encrypted project URI",      check_encrypted_channel(r)),
        ("OSPS-DO-01.01 User guides in docs",        check_user_docs(r)),
        ("OSPS-DO-02.01 Defect reporting guide",     check_defect_reporting_guide(r)),
        ("OSPS-GV-02.01 Public discussions mech.",   check_public_discussions(r)),
        ("OSPS-GV-03.01 Contribution process doc",   check_contributing_guide(r)),
        ("OSPS-LE-02.01 License meets OSI/FSF",      check_license_spdx(r)),
        ("OSPS-LE-02.02 License on release assets",  check_license_assets(r)),
        ("OSPS-LE-03.01 License file present",       check_license_file(r)),
        ("OSPS-LE-03.02 License with releases",      check_license_assets(r)),
        ("OSPS-QA-01.01 Repo publicly readable",     check_repo_public(r)),
        ("OSPS-QA-01.02 Change history public",      check_change_history(r)),
        ("OSPS-QA-02.01 Dependency list present",    check_dependency_list(r)),
        ("OSPS-QA-04.01 Subprojects list",           check_subprojects_list(r)),
        ("OSPS-QA-05.01 No binary artifacts",        check_no_binary_artifacts(r)),
        ("OSPS-VM-02.01 Security contacts doc",      check_security_contacts(r)),
    ]

    click.echo(f"\n🔒 Baseline L1 Assessment for {repo}")
    for name, result in checks:
        click.echo(f" • {name.ljust(40)} : {result}")

if __name__ == "__main__":
    main()

