requirements.txt
click
PyGithub

